// server/env.d.ts
import { z } from 'zod';
declare const ZEnv: z.ZodObject<{
    NODE_ENV: z.ZodDefault<z.ZodEnum<["development", "production", "test"]>>;
    PORT: z.ZodDefault<z.ZodNumber>;
    HOST: z.ZodDefault<z.ZodString>;
    ORIGIN: z.ZodDefault<z.ZodString>;
    LOG_LEVEL: z.ZodDefault<z.ZodEnum<["debug", "info", "warn", "error"]>>;
    REQUEST_SIZE_LIMIT: z.ZodDefault<z.ZodString>;
    RATE_LIMIT_WINDOW_MS: z.ZodDefault<z.ZodNumber>;
    RATE_LIMIT_MAX: z.ZodDefault<z.ZodNumber>;
    SUPABASE_URL: z.ZodOptional<z.ZodString>;
    NEXT_PUBLIC_SUPABASE_URL: z.ZodOptional<z.ZodString>;
    SUPABASE_SERVICE_ROLE_KEY: z.ZodOptional<z.ZodString>;
    SUPABASE_ANON_KEY: z.ZodOptional<z.ZodString>;
    NEXT_PUBLIC_SUPABASE_ANON_KEY: z.ZodOptional<z.ZodString>;
}, "strip", z.ZodTypeAny, {
    LOG_LEVEL: "info" | "error" | "warn" | "debug";
    NODE_ENV: "development" | "production" | "test";
    PORT: number;
    HOST: string;
    ORIGIN: string;
    REQUEST_SIZE_LIMIT: string;
    RATE_LIMIT_WINDOW_MS: number;
    RATE_LIMIT_MAX: number;
    SUPABASE_URL?: string | undefined;
    NEXT_PUBLIC_SUPABASE_URL?: string | undefined;
    SUPABASE_SERVICE_ROLE_KEY?: string | undefined;
    SUPABASE_ANON_KEY?: string | undefined;
    NEXT_PUBLIC_SUPABASE_ANON_KEY?: string | undefined;
}, {
    LOG_LEVEL?: "info" | "error" | "warn" | "debug" | undefined;
    NODE_ENV?: "development" | "production" | "test" | undefined;
    PORT?: number | undefined;
    HOST?: string | undefined;
    ORIGIN?: string | undefined;
    REQUEST_SIZE_LIMIT?: string | undefined;
    RATE_LIMIT_WINDOW_MS?: number | undefined;
    RATE_LIMIT_MAX?: number | undefined;
    SUPABASE_URL?: string | undefined;
    NEXT_PUBLIC_SUPABASE_URL?: string | undefined;
    SUPABASE_SERVICE_ROLE_KEY?: string | undefined;
    SUPABASE_ANON_KEY?: string | undefined;
    NEXT_PUBLIC_SUPABASE_ANON_KEY?: string | undefined;
}>;
export type Env = z.infer<typeof ZEnv>;
export declare const env: {
    LOG_LEVEL: "info" | "error" | "warn" | "debug";
    NODE_ENV: "development" | "production" | "test";
    PORT: number;
    HOST: string;
    ORIGIN: string;
    REQUEST_SIZE_LIMIT: string;
    RATE_LIMIT_WINDOW_MS: number;
    RATE_LIMIT_MAX: number;
    SUPABASE_URL?: string | undefined;
    NEXT_PUBLIC_SUPABASE_URL?: string | undefined;
    SUPABASE_SERVICE_ROLE_KEY?: string | undefined;
    SUPABASE_ANON_KEY?: string | undefined;
    NEXT_PUBLIC_SUPABASE_ANON_KEY?: string | undefined;
};
export {};